requirements.txt
# requirements.txt - Dependencies for HackRx 6.0 RAG System

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Document Processing
pdfplumber==0.9.0
pypdf==3.0.1
python-docx==1.1.0

# Embeddings and Vector Store
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# Utilities
pathlib2==2.3.7
